API updates to drivers porting from 3.10 to 4.x
This includes rewrites to a number of subsystem
drivers and their associated API's in the transition
from 3.10 to 4.x. Many of these changes will still need
further rework before being acceptable for upstream kernel
inclusion.

+++ arch/x86/include/asm/scu_ipc_rpmsg.h
@@ -2,9 +2,9 @@
 #define _SCU_IPC_RPMSG_H_
 
 struct tx_ipc_msg {
-	u32 cmd;
-	u32 sub;
-	u8 *in;
+	int cmd;
+	int sub;
+	u32 *in;
 	u32 *out;
 	u32 inlen;	/* number of bytes to be written */
 	u32 outlen;	/* number of dwords to be read */
